+++ cpp/api/include/Application.h
@@ -61,7 +61,12 @@ const int OUTPUTSTREAM = 1;
 /**
  * Option unlinked.
  */
-const int UNLINKED = 2;
+const int UNLINKED = 1 << 1;
+
+/**
+ * Option use proxy.
+ */
+const int USE_PROXY = 1 << 2;
 
 class Server;
 class EventStreamSocket;
@@ -395,10 +400,9 @@ public:
 	 * Connects to the starter application, i.e. the application which started this application.
 	 * The server and instance are returned. Be careful, the instance is linked to the server, so it must not be destroyed before.
 	 * \param options The options passed to connect the starter app.
-	 * \param useProxy True if the proxy is used to connect to the starter app.
 	 * \param timeout Timeout for the server initialization.
 	 */
-	static std::unique_ptr<ServerAndApp> connectToStarter(int options = 0, bool useProxy = false, int timeout = 0);
+	static std::unique_ptr<ServerAndApp> connectToStarter(int options = 0, int timeout = 0);
 
 	/**
 	 * Returns a string representation of this application.