+++ tests/cpp/src/TestSubscriber.cpp
@@ -41,16 +41,18 @@ int main(int argc, char *argv[]) {
 		return -1;
 	}
 
+	int options = 0;
 	bool useProxy = false;
 	string endpoint = "tcp://localhost:11000";
 	if (argc > 4) {
 		useProxy = (string(argv[3]) == "true");
 	}
 	if (useProxy) {
+		options |= USE_PROXY;
 		endpoint = "tcp://localhost:10000";
 	}
 
-	unique_ptr<Server> server = Server::create(endpoint, useProxy);
+	unique_ptr<Server> server = Server::create(endpoint, options);
 	server->init();
 
 	// Loop the number of times.