+++ tests/python/src/testappexceptions.py
@@ -6,11 +6,13 @@ this.init(sys.argv)
 
 useProxy = False if len(sys.argv) < 4 else (sys.argv[2] == "true") 
 
+options = 0
 endpoint = "tcp://localhost:11000";
 if useProxy:
+    options |= cameopy.USE_PROXY
     endpoint = "tcp://localhost:10000";
 
-server = cameopy.Server.create(endpoint, useProxy)
+server = cameopy.Server.create(endpoint, options)
 server.init()
 
 try: