+++ tests/python/src/testbasicresponder.py
@@ -16,11 +16,13 @@ if len(sys.argv) >= 4:
 
 useProxy = False if len(sys.argv) < 5 else (sys.argv[3] == "true") 
 
+options = 0
 endpoint = "tcp://localhost:11000";
 if useProxy:
+    options |= cameopy.USE_PROXY
     endpoint = "tcp://localhost:10000";
 
-server = cameopy.Server.create(endpoint, useProxy)
+server = cameopy.Server.create(endpoint, options)
 server.init()
 
 print("Responder application is", applicationName)