+++ tests/python/src/testpublishererror.py
@@ -17,11 +17,13 @@ print("Publisher application is", applicationName)
 
 useProxy = False if len(sys.argv) < 5 else (sys.argv[3] == "true") 
 
+options = 0
 endpoint = "tcp://localhost:11000";
 if useProxy:
+    options |= cameopy.USE_PROXY
     endpoint = "tcp://localhost:10000";
 
-server = cameopy.Server.create(endpoint, useProxy)
+server = cameopy.Server.create(endpoint, options)
 server.init()
 
 for i in range(numberOfTimes):