+++ tests/python/src/testresult.py
@@ -8,11 +8,13 @@ this.init(sys.argv)
 numberOfTimes = 1 if len(sys.argv) < 3 else int(sys.argv[1])
 useProxy = False if len(sys.argv) < 4 else (sys.argv[2] == "true") 
 
+options = 0
 endpoint = "tcp://localhost:11000";
 if useProxy:
+    options |= cameopy.USE_PROXY
     endpoint = "tcp://localhost:10000";
 
-server = cameopy.Server.create(endpoint, useProxy)
+server = cameopy.Server.create(endpoint, options)
 server.init()
 
 for i in range(numberOfTimes):