+++ tests/python/src/testterminate.py
@@ -20,11 +20,13 @@ this.init(sys.argv)
 
 useProxy = False if len(sys.argv) < 3 else (sys.argv[1] == "true") 
 
+options = 0
 endpoint = "tcp://localhost:11000";
 if useProxy:
+    options |= cameopy.USE_PROXY
     endpoint = "tcp://localhost:10000";
 
-server = cameopy.Server.create(endpoint, useProxy)
+server = cameopy.Server.create(endpoint, options)
 server.init()
 
 thisApp = server.connect(cameopy.This.getName())